test: Factor whitespace expansion into internal package
